debugger (debugger v5.3.2)

Erlang Debugger.

Erlang Debugger for debugging and testing of Erlang programs.

Functions

quick(Module, Name, Args)

-spec quick(Module, Name, Args) -> term() when Module :: atom(), Name :: atom(), Args :: [term()].

Debugs a single process. The module Module is interpreted and apply(Module,Name,Args) is called. This opens an Attach Process window. For details, see the User's Guide.

start/0

-spec start() -> term().

Equivalent to start/2

start/1

-spec start(Mode) -> term() when Mode :: local | global | wx;
     (File) -> term() when File :: string().

Equivalent to start/2

start(Mode, File)

-spec start(Mode, File) -> term() when Mode :: local | global, File :: string().

Starts Debugger.

If a filename is specified as argument, Debugger tries to load its settings from this file. For details about settings, see the User's Guide.

If local is specified as argument, Debugger interprets code only at the current node. If global is specified as argument, Debugger interprets code at all known nodes, this is the default.