i (debugger v5.1)

The i module provides short forms for some of the functions used by the graphical Debugger and some of the functions in module int, the Erlang interpreter.

This module also provides facilities for displaying status information about interpreted processes and break points.

It is possible to attach to interpreted processes by giving the corresponding process identity only. By default, an attachment window is displayed. Processes at other Erlang nodes can be attached manually or automatically.

By preference, these functions can be included in module shell_default. By default, they are included in that module.

See Also

int(3)

Functions

help() -> ok

Prints help text.

ia(Pid) -> ok | no_proc

Pid = pid()

Attaches to the debugged process Pid. An Attach Process window is opened for the process.

ia(Pid, Function) -> ok | no_proc

Pid = pid()

Function = {Module,Name}

Module = Name = atom()

Attaches to the debugged process Pid. The interpreter calls spawn(Module, Name, [Pid]) (and ignores the result).

ia(X,Y,Z) -> ok | no_proc

X = Y = Z = int()

Same as ia(Pid), where Pid is the result of calling the shell function pid(X,Y,Z).

ia(X,Y,Z, Function) -> ok | no_proc

X = Y = Z = int()

Function = {Module,Name}

Module = Name = atom()

Same as ia(Pid, Function), where Pid is the result of calling the shell function pid(X,Y,Z). An attached process is expected to call the unofficial function int:attached(Pid) and to be able to handle messages from the interpreter. For an example, see dbg_wx_trace.erl.

iaa(Flags) -> true

Flags = [init | break | exit]

Function = {Module,Name,Args}

Module = Name = atom()

Args = [term()]

Sets when and how to attach to a debugged process automatically, see int:auto_attach/2. Function defaults to the standard function used by Debugger.

iaa(Flags, Function) -> true

ib(Module, Line) -> ok | {error, break_exists}

Module = atom()

Line = int()

Creates a breakpoint at Line in Module.

ib(Module, Name, Arity) -> ok | {error, function_not_found}

Module = Name = atom()

Arity = int()

Creates breakpoints at the first line of every clause of function Module:Name/Arity.

iba(Module, Line, Action) -> ok

Module = atom()

Line = int()

Action = enable | disable | delete

Sets the trigger action of the breakpoint at Line in Module to Action.

ibc(Module, Line, Function) -> ok

Module = atom()

Line = int()

Function = {Module,Name}

Name = atom()

Sets the conditional test of the breakpoint at Line in Module to Function.

The conditional test is performed by calling Module:Name(Bindings), where Bindings is the current variable bindings. The function must return true (break) or false (do not break). To retrieve the value of a variable Var, use int:get_binding(Var, Bindings).

ibd(Module, Line) -> ok

Module = atom()

Line = int()

Makes the breakpoint at Line in Module inactive.

ibe(Module, Line) -> ok

Module = atom()

Line = int()

Makes the breakpoint at Line in Module active.

ic() -> ok

Clears information about processes executing interpreted code by removing all information about terminated processes.

ii(AbsModule) -> {module, Module} | error ii(AbsModules) -> ok

AbsModules = [AbsModule]

AbsModule = Module | File

Module = atom()

File = string()

Interprets the specified module(s). ii/1 interprets the module(s) only at the current node, see int:i/1. ini/1 interprets the module(s) at all known nodes, see int:ni/1.

il() -> ok

Makes a printout of all interpreted modules. Modules are printed together with the full path name of the corresponding source code file.

im() -> pid()

Starts a new graphical monitor. This is the Monitor window, the main window of Debugger. All the Debugger and interpreter functionality is accessed from the Monitor window. This window displays the status of all processes that have been or are executing interpreted modules.

ini(AbsModule) -> {module, Module} | error ini(AbsModules) -> ok

inq(AbsModule) -> ok

AbsModule = Module | File

Module = atom()

File = string()

Stops interpreting the specified module. iq/1 stops interpreting the module only at the current node. inq/1 stops interpreting the module at all known nodes.

ip() -> ok

Prints the current status of all interpreted processes.

ipb() -> ok

Prints all existing breakpoints.

ipb(Module) -> ok

Module = atom()

Prints all existing breakpoints in Module.

iq(AbsModule) -> ok

ir() -> ok

Deletes all breakpoints.

ir(Module) -> ok

Module = atom()

Deletes all breakpoints in Module.

ir(Module, Line) -> ok

Module = atom()

Line = int()

Deletes the breakpoint at Line in Module.

ir(Module, Name, Arity) -> ok | {error, function_not_found}

Module = Name = atom()

Arity = int()

Deletes the breakpoints at the first line of every clause of function Module:Name/Arity.

ist(Flag) -> true

Flag = all | no_tail | false

Sets how to save call frames in the stack, see int:stack_trace/1.

iv() -> atom()

Returns the current version number of the interpreter. Same as the version number of the Debugger application.